requirements.txt
python-docx==1.2.0
numpy==2.4.0
openai==2.14.0
faiss-cpu==1.13.2
mistralai==1.10.0
python-dotenv==1.2.1
tenacity==9.1.2
rich==14.2.0
